docker build
ajout du fichier ".dockerignore" : comme pour .git les fichiers listés ici ne sont pas copiés par l'instruction COPY . du Dockerfile
copier "install.sh" dans /root car on l'utilise pendant la séquence d'initialisation "init.sh"
Ajout du "healthcheck": permet de valider la bonne initialisation du container, on peut ainsi le tester dans le workflow
Ne pas re-télécharger le source jeedom durant l'initialisation (init.sh) puisqu'on l'a déjà fait durant le build (Dockerfile).
Fix la fonction "trap" à l'arrêt du conteneur

+++ install/OS_specific/Docker/init.sh
@@ -81,10 +81,16 @@ if [ -f ${WEBSERVER_HOME}/core/config/common.config.php ]; then
 else
 	echo 'Start jeedom installation'
 	JEEDOM_INSTALL=0
-	rm -rf /root/install.sh
-	wget https://raw.githubusercontent.com/jeedom/core/${VERSION}/install/install.sh -O /root/install.sh
-	chmod +x /root/install.sh
-	/root/install.sh -s 6 -v ${VERSION} -w ${WEBSERVER_HOME}
+
+  # do not re-install jeedom
+  if [ ! -f ${WEBSERVER_HOME}/core/config/common.config.sample.php ]; then
+    echo 'download again Jeedom'
+    rm -rf /root/install.sh
+    wget https://raw.githubusercontent.com/jeedom/core/${VERSION}/install/install.sh -O /root/install.sh
+    chmod +x /root/install.sh
+    /root/install.sh -s 6 -v ${VERSION} -w ${WEBSERVER_HOME}
+  fi
+
 	if [ $(which mysqld | wc -l) -ne 0 ]; then
 		chown -R mysql:mysql /var/lib/mysql
 		mysql_install_db --user=mysql --basedir=/usr/ --ldata=/var/lib/mysql/
@@ -101,7 +107,9 @@ else
 		sed -i "s/#USERNAME#/jeedom/g" ${WEBSERVER_HOME}/core/config/common.config.php
 		sed -i "s/#PORT#/3306/g" ${WEBSERVER_HOME}/core/config/common.config.php
 		sed -i "s/#HOST#/localhost/g" ${WEBSERVER_HOME}/core/config/common.config.php
+    # jeedom installation : install composer, init databse
 		/root/install.sh -s 10 -v ${VERSION} -w ${WEBSERVER_HOME}
+    # jeedom post-install
 		/root/install.sh -s 11 -v ${VERSION} -w ${WEBSERVER_HOME}
 	fi
 fi
@@ -148,4 +156,9 @@ service apache2 start
 echo "Add trap docker_stop"
 trap "docker_stop $$ ;" 15
 
-cron -f
+# launch cron daemon in background
+cron -f &
+# hold the cron daemon PID
+child=$!
+# wait for cron stopping
+wait "$child"
